ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
Pillow==11.1.0
torch==2.5.1
torchvision==0.20.1
git+https://github.com/openai/CLIP.git
faiss-cpu==1.9.0
sqlalchemy==2.0.36
mysql-connector-python==9.1.0
python-multipart==0.0.18
requests==2.32.3
numpy==1.26.4
